requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai
openai
pypdf
chromadb
tiktoken
unstructured
